requirements.txt
bson==0.5.10
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
h11==0.12.0
idna==3.1
motor==2.4.0
pydantic==1.8.1
pymongo==3.11.3
python-dateutil==2.8.1
python-decouple==3.4
six==1.15.0
starlette==0.13.6
style==1.1.0
typing-extensions==3.7.4.3
update==0.0.1
uvicorn==0.13.4
pytz